separated out skebs to credit/skebs.py, addressing reviewer comments, enabled CI tests
