+++ tests/test_physics.py
@@ -5,19 +5,13 @@ import os
 import xarray as xr
 
 import torch
-from credit.physics_core import compute_pressure_on_mlevs
+from credit.physics_core import ModelLevelPressures
 
 TEST_FILE_DIR = "/".join(os.path.abspath(__file__).split("/")[:-1])
 CONFIG_FILE_DIR = os.path.join("/".join(os.path.abspath(__file__).split("/")[:-2]),
                       "config")
 
 def test_pressure_on_mlevs():
-    # testing on derecho
-    # level_info = xr.open_dataset("/glade/u/home/wchapman/MLWPS/DataLoader/ERA5_Lev_Info.nc")
-    # era5_ds = xr.open_zarr("/glade/derecho/scratch/wchapman/SixHourly_y_TOTAL_2020-01-01_2020-12-31_staged.zarr")
-    # a_vals = level_info["a_model"].sel(level=era5_ds["level"])
-    # b_vals = level_info["b_model"].sel(level=era5_ds["level"])
-
     # test on gh actions
     level_info = xr.open_dataset(os.path.join(TEST_FILE_DIR, "data/level_info_test.nc"))
     a_vals, b_vals = level_info["a_model"], level_info["b_model"]
@@ -34,7 +28,7 @@ def test_pressure_on_mlevs():
 
     a_tensor = a_tensor.view(1, levels, 1, 1, 1)
     b_tensor = b_tensor.view(1, levels, 1, 1, 1)
-    plev_converter = compute_pressure_on_mlevs(a_tensor,
+    plev_converter = ModelLevelPressures(a_tensor,
                                                b_tensor,
                                                plev_dim=1)
     thickness = plev_converter.compute_mlev_thickness(y_pred[:, sp_index: sp_index + 1])
