+++ tests/test_postblock.py
@@ -3,87 +3,30 @@
 -------------------------------------------------------
 Content:
 """
-import pytest
 import yaml
 import os
 import logging
 
 import torch
-import torch.nn as nn
-from credit.models.crossformer import CrossFormer
-from credit.postblock import GlobalWaterFixer, PostBlock, Backscatter_FCNN
-from credit.postblock import SKEBS, TracerFixer, GlobalMassFixer, GlobalEnergyFixer
+from credit.postblock import GlobalWaterFixer, PostBlock
+from credit.skebs import SKEBS, BackscatterFCNN
+from credit.postblock import TracerFixer, GlobalMassFixer, GlobalEnergyFixer
 from credit.parser import credit_main_parser
-from credit.postblock import Backscatter_CNN
 
 
 TEST_FILE_DIR = "/".join(os.path.abspath(__file__).split("/")[:-1])
 CONFIG_FILE_DIR = os.path.join("/".join(os.path.abspath(__file__).split("/")[:-2]),
                       "config")
 
-@pytest.mark.skip(reason="need to have model weights, to test")
-def test_SKEBS_integration():
-    """
-    integration testing to make sure everything goes on GPU, is loaded properly etc
-    requires loading weights
-    """
-    logging.info("integration testing SKEBS")
-    # config = os.path.join(CONFIG_FILE_DIR, "example_skebs.yml")
-    config = "/glade/work/dkimpara/CREDIT_runs/latest_skebs/latest_skebs.yml"
-    with open(config) as cf:
-        conf = yaml.load(cf, Loader=yaml.FullLoader)
-
-    conf = credit_main_parser(conf) # parser will copy model configs to post_conf
-
-    post_conf = conf['model']['post_conf']
-    ## setting up input
-    image_height = post_conf["model"]["image_height"]
-    image_width = post_conf["model"]["image_width"]
-    channels = post_conf["model"]["channels"]
-    levels = post_conf["model"]["levels"]
-    surface_channels = post_conf["model"]["surface_channels"]
-    output_only_channels = post_conf["model"]["output_only_channels"]
-    input_only_channels = post_conf["model"]["input_only_channels"]
-    frames = post_conf["model"]["frames"]
-    sp_index = post_conf["skebs"]["SP_ind"]
-
-    in_channels = channels * levels + surface_channels + input_only_channels
-    x = torch.randn(2, in_channels, frames, image_height, image_width)
-    out_channels = channels * levels + surface_channels + output_only_channels
-    y_pred = torch.randn(2, out_channels, frames, image_height, image_width)
-    y_pred[:, sp_index] = torch.ones_like(y_pred[:, sp_index]) * 1013
-
-    model = CrossFormer(**conf["model"])
-    device = torch.device(f"cuda:{1 % torch.cuda.device_count()}") if torch.cuda.is_available() else torch.device("cpu")
-    model = model.load_model(conf)
-    model.to(device)
-    model.train()
-    pred = model(x.to(device))
-
-    # model = SKEBS(conf["model"]["post_conf"])
-    # device = torch.device(f"cuda:{1 % torch.cuda.device_count()}") if torch.cuda.is_available() else torch.device("cpu")
-    # model.to(device)
-    # model.train()
-    # pred = model({"y_pred": y_pred.to(device)})
-
-    loss_fn = nn.MSELoss()
-    loss = loss_fn(pred.float(), y_pred.to(device).float()).mean()
-    # dot = make_dot(loss, params=dict(model.named_parameters()))
-    # dot.format = 'png'
-    # dot.render('model_arch.png')
-    scaler = torch.GradScaler()
-    scaler.scale(loss / 1).backward()
-
-    assert pred.shape == y_pred.shape
-
-@pytest.mark.skip(reason="need to have scaler file and static file to test")
 def test_SKEBS_rand():
-    ''' unit test for CPU. testing that values make sense
-    '''
+    """ unit test for CPU. testing that values make sense
+    """
     config = os.path.join(CONFIG_FILE_DIR, "example-v2025.2.0.yml")
     with open(config) as cf:
         conf = yaml.load(cf, Loader=yaml.FullLoader)
 
+    conf["data"]["save_loc_static"] = os.path.join(TEST_FILE_DIR, "data/level_info_test.nc")
+    conf["data"]["mean_path"] = os.path.join(TEST_FILE_DIR, "data/mean_6h_1979_2018_16lev_0.25deg.nc")
     conf['model']['post_conf']["activate"] = True
 
     conf['model']["post_conf"]["global_mass_fixer"] = {"activate": False}
@@ -92,11 +35,18 @@ def test_SKEBS_rand():
     conf['model']["post_conf"]["tracer_fixer"] = {"activate": False}
 
     conf['model']["post_conf"]["skebs"]["activate"] = True
+    conf['model']["post_conf"]["skebs"]["dissipation_type"] = "uniform"
+
+    image_height = 640 # needs to match level_info_test.nc
+    image_width = 1280
+
+    conf["model"]["image_height"] = image_height
+    conf["model"]["image_width"] = image_width
+
     conf = credit_main_parser(conf) # parser will copy model configs to post_conf
     post_conf = conf['model']['post_conf']
     
-    image_height = post_conf["model"]["image_height"]
-    image_width = post_conf["model"]["image_width"]
+   
     channels = post_conf["model"]["channels"]
     levels = post_conf["model"]["levels"]
     surface_channels = post_conf["model"]["surface_channels"]
@@ -119,7 +69,6 @@ def test_SKEBS_rand():
 
     skebs_pred = postblock(input_dict)
 
-    # FIXME: fix test
     assert skebs_pred.shape == y_pred.shape
     assert not torch.isnan(skebs_pred).any()
 
@@ -142,7 +91,7 @@ def test_SKEBS_backscatter():
     out_channels = channels * levels + surface_channels + output_only_channels
     y_pred = torch.randn(2, out_channels, frames, image_height, image_width)
 
-    model = Backscatter_FCNN(out_channels, levels)
+    model = BackscatterFCNN(out_channels, levels)
 
     pred = model(y_pred)
 
@@ -150,23 +99,6 @@ def test_SKEBS_backscatter():
     target_shape[1] = levels
     assert list(pred.shape) == target_shape
     assert not torch.isnan(pred).any()
-
-def test_backscatter_pad():
-    lat = 2
-    lon = 2
-    batch_size = 1
-    x = torch.arange(batch_size * lat * lon).reshape(batch_size, 1, lat, lon)
-    m = Backscatter_CNN(1,1,lat,lon)
-    
-    target = torch.tensor([[[[0, 1, 0, 1],
-                            [1, 0, 1, 0],
-                            [3, 2, 3, 2],
-                            [2, 3, 2, 3]]]])
-    
-    padded = m.pad(x)
-    assert (padded == target).all()
-
-    assert (m.unpad(padded) == x).all()
 
 def test_TracerFixer_rand():
     """Provides an I/O size test on TracerFixer at credit.postblock."""
